requirements.txt
alabaster==0.7.12
Babel==2.9.1
biopython==1.79
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
decorator==4.4.2
docutils==0.17.1
filelock==3.0.12
Flask==2.0.1
funcsigs==1.0.2
gdown==3.13.0
GridDataFormats==0.5.0
gsd==2.4.2
gunicorn==20.1.0
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.4
imagesize==1.2.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
latexcodec==2.0.1
lxml==4.6.3
MarkupSafe==2.0.1
MDAnalysis==1.1.1
mmtf-python==1.1.2
mock==4.0.3
moviepy==1.0.3
msgpack==1.0.2
networkx==2.5.1
notebook==5.7.10
numpy==1.21.2
opencv-python==4.5.3.56
packaging==20.9
pandas==1.3.2
Pillow==8.3.1
proglog==0.1.9
pybtex==0.24.0
pybtex-docutils==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
sk-video==1.1.10
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.3.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
torch==1.4.0
torchvision==0.5.0
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
